fix: update test mocks for f64 temperature and ClusterConfig.cluster_id
- ReactiveHandler type alias: f32 → f64 to match LLMProvider trait
- Add missing cluster_id field to ClusterConfig initializers in cluster tests

+++ tests/live/mocks/reactive_llm.rs
@@ -13,7 +13,7 @@ use bendclaw::llm::tool::ToolSchema;
 use bendclaw::llm::usage::TokenUsage;
 
 type ReactiveHandler =
-    dyn Fn(usize, &[ChatMessage], &[ToolSchema], f32) -> Result<LLMResponse> + Send + Sync;
+    dyn Fn(usize, &[ChatMessage], &[ToolSchema], f64) -> Result<LLMResponse> + Send + Sync;
 
 fn mock_usage() -> TokenUsage {
     TokenUsage {
@@ -32,7 +32,7 @@ pub struct ReactiveMockLLMProvider {
 
 impl ReactiveMockLLMProvider {
     pub fn new(
-        handler: impl Fn(usize, &[ChatMessage], &[ToolSchema], f32) -> Result<LLMResponse>
+        handler: impl Fn(usize, &[ChatMessage], &[ToolSchema], f64) -> Result<LLMResponse>
             + Send
             + Sync
             + 'static,
